Move `Turbo` window declaration to `index.ts`
