+++ src/index.ts
@@ -1,4 +1,6 @@
+import * as Turbo from "@hotwired/turbo"
 import { TurboStreamAction, TurboStreamActions } from "@hotwired/turbo"
+import { BrowserAdapter } from "@hotwired/turbo/dist/types/core/native/browser_adapter"
 
 import * as TurboMorph from "turbo-morph"
 import * as Actions from "./actions"
@@ -6,6 +8,16 @@ import * as Utils from "./utils"
 
 export * as Actions from "./actions"
 export * as Utils from "./utils"
+
+declare global {
+  interface Window {
+    Turbo: typeof Turbo & {
+      navigator: {
+        adapter: BrowserAdapter
+      }
+    }
+  }
+}
 
 export function initialize(streamActions: TurboStreamActions) {
   TurboMorph.initialize(streamActions)
